Bug #5284 -- improve module autoloading documentation.
Import a relatively clear example from the mailing list, and generally try to
improve the module autoloading documentation.

That means I restructure the text to put the emphasis on the things we want to
happen rather than the things that are undesirable but possible.  Those get
left for a tiny little tip so they look less important to folks.