(#6359) Fix spectacular formatting issues in all references prior to 2.6.0
Hey, it turns out that spitting out ReST of questionable validity and then funneling it through Pandoc before processing the resultant file with Jekyll results in some pretty gnarly HTML!

This patch fixes all that, hopefully. Basically, I got the original RST for each version, converted it to html with rst2html-2.6.py, did a series of increasingly absurd s///g's to clean out the garbage, then converted the headers into markdown (for Jekyll's indexing) and left all the rest of the HTML as-is. The great thing about Markdown is that this kind of laziness still results in a perfectly valid file.

Needless to say, please don't re-generate these files again unless there's a critical need.