Improve stack of monospaced fonts in style.css
The Puppet DSL has chaining operators with dashes and tildes. The two versions have different effects, but are used in the same places. The Courier and Courier New tilde and dash symbols are almost indistinguishable at sizes smaller than approximately 20pt. We can't keep using the default monospace font for code snippets, because 9 times out of 10 it's Courier or Courier New.