Make all pages in /mcollective/ display a custom sidebar instead of a page TOC
This commit adds a simple filter for getting the top-level directory of a page,
an include snippet containing R.I.'s preferred sidebar HTML, and an if statement
in default.html that swaps in that sidebar when the current page is inside the mcollective docs.
It should not affect any future documents whose names include "mcollective" but are not
inside the main mcollective directory.