(#6891) Remove meta-syntactic variables.
Foo, bar, baz are potentially confusing  when used in documentation.
Particularly examples.  They're evil because the author may have a clear
idea of what legal substitutions are in their place, but the reader
likely does not have this experience or knowledge.

This commit tries to replace each instance of the variables with much
more concrete and common examples.  Apache, NTP, message of the day,
etc...

In addition, the commit adds a bit more explanation for the tests
directory.