Shrink font size in td and th elements to 75% of what they inherit.
Notice how the PDF text is squashed to the left a bit? By like an inch? It's because of a wide table in the user resource reference. I already dealt with the worst of that with th_underscores.js, but that didn't totally fix it.

I don't know how to do that for the long term, actually; some tables are just going to be wide, and we need to keep them from deforming the page. In the meantime, this should solve it at the expense of making table text smaller. There are like ten tables in a 280-pg document, I'm not too fussed about it.