Tweak site CSS to modernize TOC hierarchy
The CSS was setting list-style to none on the sidebar OLs and then generating numbers with
a li:before by keeping track of three separate counter variables. Also, it would only
work with a specially crafted set of nested OL and UL elements, i.e. if you want an
ordered list you needed to ask for an unordered one. Insanity. Anyway, I'm killing that
and using normal list numbering. Call me if it breaks anything. Thx to Randall for CSS help.