Add class parameter defaults recommendation
This recommendation and associated SHOULD statements result from an
issue filed against the puppetlabs-mysql module [1]

The issue is that a class parameter can only pull variables from the
class it inherits in Puppet version 2.6.3 and later.  Puppet version
2.6.2 and earlier result in a parse error.

This recommendation is aimed at maximizing the usability of reusable
Puppet modules across 2.6.x versions.

[1] https://github.com/puppetlabs/puppetlabs-mysql/issues/3