(#8140) Update fog validation instructions
This change updates the fog validation instructions for the getting
started with CloudPack guide.  If Fog is properly configured, but no
servers have been booted in EC2, the test will return 0.

This change updates the conditional to be >= 0 rather than > 0 and adds
additional information about what an improperly configured ~/.fog file
looks like.