(#8520) Change body ID from "resources" to "puppetlabsdocs"
A poorly chosen (reason: overly common) ID for the body element is
causing persistent problems in the type reference, as no one can link to
the "resources" type. This commit changes the stylesheet, the default
layout, and the corresponding layout for generating PDFs in order to
avoid this problem.

As docs site CSS appears to have diverged permanently from the main site
CSS, I'm not going to spend much time worrying about where else body
id="resources" is being used.