Maint: Chop down insane CSS
Since the docs writers don't have access to the original Dreamweaver
source (or whatever abstraction layer was used to generate this
monster), the 4000 lines of mostly irrelevant rules in style.css have
occasionally been a demoralizing barrier to maintenance. This commit
removes a large number of rules that could not possibly apply anywhere
in the one template the docs project uses.