Mention requirement of loadall call when trying to call function from function
Zach ran into this -- when calling a function from a function, it will
sometimes just fail completely, and adding this line...

Puppet::Parser::Functions.autoloader.loadall

...seems to fix it. This may have something to do with issue #5959, which was
where I found it in the first place. Anyway, to the best of my knowledge it's
voodoo, but since it appears to be necessary, I'm throwing it in the docs.