Edit 2.7.1 service type reference to match edits made to 2.7.0
Due to 2.7's change to hasstatus's default value, I had to make a
post-generation edit to the service type's documentation in the 2.7.0 docs.
(see puppet-docs commit 8270c2e.) This commit edits the 2.7.1 docs to match;
further edits won't be necessary, as the docs were fixed in the puppet code
for 2.7.3.