CSS: Make images alone in paragraphs be centered
Actually, I'm a little shaky on this one -- it might also turn inline images
in a text graf into blocks, as long as the <p> element doesn't contain any <a>
or <em> elements or whatever, which would seem non-deterministic and bizarre
in practice. But I don't think we should be using images inline in text
in the first place, so hopefully that won't come up.