Fix issue where links into directories with periods in them weren't being translated
We were properly gsubbing dots for dashes in the filename part of hrefs, but
we weren't doing the same for the directory portion of the path. This was
causing a problem for the dashboard and pe 1.2 manuals. This commit gsubs
periods across the board, although it still looks away from the possibility of
folders with #s or whatever in their names. (Don't do that.)