Baselines: Mark the standalone baseline docs as deprecated
It came to my attention that the customer-specific baseline plugin docs were
being indexed by Google, and that at least one PE customer was attempting to
learn about the compliance features by reading these non-PE docs.

We can't take the doc offline, as it was part of the deliverable to the plugin
customer; however, this commit adds a (hopefully unignorably lurid)
deprecation warning and redirects most users to the PE manual.