PE: Remove confusing comment in pe_1.2_nav from final output
The commented section of _layouts/pe_1.2_nav.markdown was being carried over
to the final site, there to confuse anyone who checked the page source. This
commit switches it to a Liquid comment, which is removed before the site is
rendered to HTML. It also changes the source of the navigation to use Markdown.