PE2: Convert navigation to a sidebar with external document hierarchy
This commit adds a new layout, for use with the PE 2.0 manual ONLY, which
changes the navigation to sidebar. This is a prototype for future
directions in the docs' approach to navigation.

The new sidebar layout displays the entire TOC of the pe2 manual, and
inserts the in-page TOC into the appropriate place in the hierarchy, highlighted
in orange.