PE2: Improve navigation by collapsing major sections
This commit extends the experimental nav code so it can collapse major sections.
I went the easy route with expansion: Clicking a collapsed section takes you to
the first page in that section, and we don't have per-section landing pages.

Needless to say, the code for this prototype is hairy and scary.