PE2: Make experimental nav sidebar highlight sections instead of individual chapters
I was having trouble figuring out which link was my "next" one in the nav sidebar.
This commit tries to address that by highlighting the current _section_ instead
of the current in-page anchors. It also lightens the in-page anchors (signifying
that they've been "visited") so they don't draw clicks away from that crucial
next page.