(#11072) Development lifecycle inline wiki content
http://docs.puppetlabs.com/guides/development_lifecycle.html links to a
wiki page
http://projects.puppetlabs.com/projects/puppet/wiki/Development_Lifecycle
that has little info and just links to github. The wiki page used to
have all the info about how to contribute to Puppet and changed fairly
frequently, but now that the info has been moved out it shouldn’t change
often at all and should just be part of the docs site. The wiki page
will need to stay as a place holder also since not to long ago that’s
where we told people to look for dev info, but no reason to have people
follow a frustrating set of links just to get to the info.