(#11168) Mention Puppet Enterprise in the installation guide
Even though there's a banner on the docs site for Puppet Enterprise, it
should be mentioned when people install puppet from the installation
guide, especially since that's the first page that you hit when you
Google 'install puppet'.

I tried to mention Puppet Enterprise without being tacky about pushing
the commercial product, and linked directly to the comparison and FAQ so
that it was clear why you might choose to pay for Puppet Enterprise or
not.