(#11267) Update FAQ to state Puppet does run on Windows
While Googling for ‘puppet windows’ the FAQ entry was 3rd in the results and
says we don’t yet have windows support, despite the Wiki with all the great
info on how to use it on windows. This is now updated to reflect the current
state of affairs and link to the Wiki.