(maint) Add link to RDoc markup document in Style Guide
Without this patch the style guide gives an example of the inline puppet
manifest documentation but does not mention the markup format is RDoc.

This patch makes it clear module authors should specifically write
inline documentation in RDoc and provides a reference link to where to
learn about the syntax of the RDoc markup.

It also makes it clear why this is important; to automatically generate
documentation using puppet doc.