(#11312) Reword modules section intro
Previous to this commit, the first words
in the modules section is 'So anyway,'.
Certain parts of the site link directly
to this section using a page anchor.
Users that follow this link will be
confused and feel like they've missed
important context further on the page.

This commit removes the words entirely