PDF: Make PE 2.0 pdf theoretically work
The unwholesome and vile abuse of Jekyll I used to get a hierarchical TOC
sidebar for PE2 broke the unwholesome and vile abuse of Jekyll I used to get
our PDFs. Adding a ringer pe2experimental layout identical to the ringer
default layout should allow all unwholesome and vile abuses to live in
harmony.

Q: Why didn't you just symlink it?
A: Because Jekyll will work with directory symlinks, but will blow up if asked
to read from a symlink to a plain file.

Q: Why do you know that?
A: Do not ask.