Add a header section for the curated guides
The primary purpose of this change is to provide a specific target for the
list of non-automatically-generated docs, so it pops out better in
the table of contents. Right now the TOC has a list of items
under "Reference Shelf" but the only option in the list is "Generated
References" so it's not obvious that there are other options.