LP: Add VM-configuring "learningpuppet" module
Readers will need an agent VM, and the quickest way to get one seems to be
duplicating the already-downloaded learning puppet VM. This module will be
offered as a quick way to do the really fiddly bits of re-configuring a puppet
master VM to act like a just-born agent VM. See the documentation in the
module for more details.

This module only works with a VM running PE 2.0.x.