PE: Mention that console's cert settings may be missing when upgrading from 1.1 (#11464)
This commit adds a list of cert-related settings to the upgrade guide, and mentions
that they may be missing entirely from settings.yml depending on the PE version
from which you are upgrading. It also adds a short troubleshooting note in the step
where the absence of these settings would cause failures.

This information was added to the PE 1.2 and 2.0 guides.