PE2: Fix slight inaccuracy in database name/user answers
Per feedback from Matthaus: "Both of these answers are used regardless of how
the database is being handled. If it is being automatically configured, they
are used when databases and users are created. If it is not being
automatically configured, they are used to connect and in settings files
(puppet.conf and databases.yml)."