Remove guides/installing_dashboard.markdown
This one needs a 301 on the server, to dashboard/manual/1.2/bootstrapping.html.

We removed all links to this deprecated guide, but I forgot about Dre^H^H^H the
sitemap, which left it very VERY visible to Google and therefore anyone searching
for "install puppet dashboard." So now we kill it with fire.