Guts: Add 'iflink' conditional link tag
This commit adds an "iflink" tag.

Usage:

{% iflink "This is the text of the link", "./relative_path.html" %}
{% iflink "Text of another link", "http://example.com/absolute/path/with_or_without_protocol_and_hostname.html" %}

Output:

If the URL is the current page:
<span class="currentpage">This is the text of the link</span>

If the URL isn't the current page:
<a href="./relative_path.html">This is the text of the link</a>

As suggested by the examples, it takes the following types of path as its URL parameter:

* ./relative.html
* /absolute/path.html
* https://example.com/absolute/path.html

Don't use ../, that would be silly. ALSO, note that DOUBLE QUOTES ARE NON-OPTIONAL.
No single-quotes allowed. The comma between arguments is required, but the
space after the comma is optional.