Guts: Take build phase of deployment out of vlad
Since the build phase is just building a local tarball, there's no reason to
do it in the vlad namespace (which requires that you call the mirror0 task or
something before invoking build; it was weird). So this commit moves it into
the main rakefile and changes the references.

This commit also removes the redundant vlad:deploy task, which appeared to be
mostly identical to the normal deploy task.