CSS: Restyle blockquotes so we can use them asides
This commit unstyles blockquotes so they don't look like quotes. This lets us
use them as asides instead of quotes.

Yes, this violates semantic html5-ish goodness (we should properly be using
<aside> elements). Unfortunately:

* It's really convenient to make a Markdown blockquote.
* We don't use blockquotes for actual QUOTES anywhere in the docs, which
  effectively makes it an abandoned element.
* The Markdown spec, such as it is, says processors should ignore any further
  Markdown formatting inside an explicit HTML block element (like <aside>).

You can turn the behavior in point 3 off in some processors, but I would
prefer to not rely that heavily on implementation details, and I would REALLY
prefer to not spam raw HTML in our Markdown documents. So blockquotes it is.

We may revisit this decision later.