Guts: Keep the name of the laptop generating the refs out of the configuration reference
The default value for certname isn't, in fact, magpie.puppetlabs.lan, nor was
it pelin.lovedthanlost.net. (Likewise with reportfrom.) The leakage is
unavoidable because defaults.rb has to get an original value $certname
somehow, so this commit just does a gsub after the fact when we're generating
website-ready references.