Put example machines on the example.com domain
Let's follow RFC 2606! This commit changes any .magpie.lan or .puppetlabs.lan
hostnames to be on example.com instead, and generally tightens up the
less-comprehensible machine names. ("Redmaster" was an actual test VM whose
name made sense in context, but one has to admit it sounds sinister on its own.)

We are still using bird names for most of the example nodes, because of reasons.