(#13057) Link to Facter no longer exists.
Prior to this commit, under Ruby Prerequisites in the installation
guide, the link to Facter was a dead link. I've changed it to the
closet thing I can find which is http://puppetlabs.com/puppet/related-projects/facter/
That said, that's an absolutely terrible page on Facter (containing
links to reductivelabs.com) and should be replaced when the FOSS
website launches.