Guts: Try to fix failures when deploying to second mirror
Rake deploy was failing to update the second mirror, and we needed an
additional rake mirror2 vlad:release command to get to a good state. As far as
I could tell, vlad's tools were connecting back to mirror0 to do the
rsync/untar/rsync-into-place.

I'm speculating that this is because we're calling #reenable on the
vlad:release task while the variables are still set to mirror0; if that's the
case, reversing the calls should do the trick. If not, we'll have to go back
to shelling out for those tasks, which is annoying but whatever.