Move page subtitle above the in-page index
Per Randall's suggestion, the page title should be above the page index. But
the page title displayed outside the header material was previously just done
with a H1 element, and it's actually more of a subtitle than anything else --
the page title tends to show the context within the larger docs structure, and
the subtitle just gets to the immediate point.

Anyway, this commit adds a new optional subtitle key in the page yaml, and if
it isn't there, we just use the page title instead. And it puts it above the
index. (In the pe2experimental layout.)