Un-break pages still using the default layout
This commit brings back a copy of style.css from before I started messing with
it as style_legacy.css, and switches the default layout to use that instead of
the modified one. This seemed cheaper time-wise than trying to make them
co-exist, especially since the default layout is slated for replacement
soon anyhow.