JS: Fix drop-down menus
Previously, the drop-down menus in the header would disappear as soon as you
moused out of them. This was abusive to users who orbit controls before
committing to them, and it had an especially nasty effect when combined with a
long page title in the docs. (This would cause a little gap between the quick
nav control and its menu, and crossing the chasm to reach the links would
cause the menu to disappear. Troll menu.)

I got Pieter van de Bruggen to rewrite the dropdown javascript, and now they
disappear when you click elsewhere. Fidgety users unite!