Guts: Remove <lastmod> from sitemap generator
The lastmod was just statting the source files, which is horrible because they
live in git. To have an accurate lastmod value, we'd have to do a git log on
each one. But Nigel tells us we can probably just delete it instead and rely
on Google to figure it out. Let's try that.

This commit also adds a Silent Hill joke to the same file.