Mention that PL offers native packages
We offer native packages for apt, yum, mac and windows on our own hosts
that are kept updated.  This change is from a puppet-users post that
mentioned that squeeze had 2.6.2 and was quite outdated.  I mentioned
apt.puppetlabs.com and he said it was missing from this guide.  I have
attempted to rectify that situation.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>