Add 10 most recent commits to front page of docs site
It occurred to me that the docs have a habit of silently improving -- unless
you know to watch the puppet-docs repo, there's not much transparency. Let's
change that! We'll expose the 10 most recent commits on the docs front page.
Only on the front page, though, so it isn't irritating.

We'll also include a link to the full list of commits and a feed you can follow.

This may result in a downtick in the number of whimsical commit messages;
we'll see.