Facter: Remove bogus "Facter 1.7.0" info from custom facts guide
This info was added when it looked like we were going to have a 1.7 release of
facter (instead of jumping to 2.x) very very soon. Instead, the external facts
stuff got pushed to 2.1.x, and will be working somewhat differently. Until we
are able to rehab the docs, this commit removes them from the net.