Add bits about keytool to 2.5 basic install virtual-java section.
If the keytool binary isn't in root's PATH, the pe_mcollective module will fail
to set up the collective on the master side. This adds a bit of documentation
to the basic install section to explain how to find keytool and link it into
the path if it isn't there already.