Kill "network" reference, at least for 2.7.x
This was in an unknown state for a while, but Patrick and I confirmed today
that the XMLRPC handlers (the network endpoints documented in
/references/<version>/network.html, and the namespaceauth.conf file that
controlled access) are dead code paths in the 2.7.x series. They are almost
certainly dead paths in 2.6.x too, but I need to do more testing.

Regardless, I think it's safe to at least stop generating the reference for
2.6, and we can definitely remove all trace of the reference in 2.7. This
commit does so.