PRM: Fix incorrect documentation of allowed tag names
In point of fact, Puppet downcases and de-dupes tag names when compiling the
catalog. Any mixed-case dupes will show up in the value of the resource's
"tag" parameter in the catalog, but the separate "tags" key for the resource
will have the cleaned-up list. Also, the main resource-stashing backends
expect lowercase tag names, and the next version of PuppetDB will enforce them
in its search interface.