Guts: Add jump-to-newest-counterpart links to outdated PE pages; add custom 404
Since we sometimes change file names when we cut new versions of large
documents, the jump links will be convenient most of the time and broken some
of the time. This commit adds a more helpful 404 page that explains the likely
problem and links directly to the TOC of the version you were trying to get
to.

Note that we need to add the following to the docs site vhost, in order to
activate the new 404:

  <LocationMatch ^/(pe|puppet|puppetdb|facter)/>
    # For sections of the site that contain different versions, pages may appear
    # and disappear between versions. Give a helpful error page when that happens.
    ErrorDocument 404 /files/errors/versioned_404.html
  </LocationMatch>