Jekyll: Patch our copy to allow symlinked directories
We use symlinked directories to bring in foreign content from
MCollective (and possibly other projects in the future), and our
environment is set up such that this isn't a legitimate security
concern. Protecting us from symlinks is unnecessary, and would require a
more complicated site-generation workflow. As such, we'll just patch Jekyll.