Guts: Make conditional link ("iflink") tags handle bare directory links properly
Links to /something/something/ go to /something/something/index.html,
but Apache handles that and Jekyll doesn't know it; page.url always has
the full filename. This commit teaches the conditional link tag about
index.html so we can link to /something/something/ in our sidebar pages.
Still can't link to /something/something and expect it to recognize it,
though.