Multi-master: De-emphasize SRV records, and flag them as an experimental feature
This commit re-orders the options for distributing agent load and CA
traffic to put SRV records at the bottom. They're cool and they work,
but they have some pitfalls the user needs to know about, and the
feature could use some love before we can call it the preferred method.