Update 1.1 docs to include info about PuppetDB module
These changes came (roughly) from the README in the puppetdb source
tree; they add information about the option of installing and
configuring puppetdb from the PuppetDB module, in addition to the
other documented options.  The "install_via_module.markdown" will
probably need to be edited to directly include information from
the PuppetDB module's README_GETTING_STARTED.md, rather than
just linking to it, but I just wanted to get on the same page
with the docs team about the general organization of this before
putting more time into that.