Add/modify canonical links (set to /1/...)
This commit adds canonical links to the PuppetDB 1 and 1.1 pages. All
canonical links are currently set to point to the 1.0 manual, and will
be switched later.

Some of the canonical links for 1.1 are currently nonsensical, since
they use the page's current path, which often doesn't exist in the 1.0
manual. This will be fixed later.