(#19294) Allow double quotes for strings when string contains single quotes
Basically, single quotes within a string need to be escaped when the string is
wrapped in single quotes, thus making the overall line harder to read, breaking
philosophy #1.

The tooling obviously throws an error when you try to use double quotes here,
but without clear guidance one just ends up in a debate. My patch is about being
clearer for this edge-case, for the sake of philosophy #1 (readability).

Signed-off-by: Ken Barber <ken@bob.sh>