Copying first page of deploy guide to index.markdown
This simplifies the URL structure, so readers can go to /guides/deployment_guide and find content without knowing a specific page name.

Concurrent with this change, I'm submitting a pull request to have the first page 301'd to index.html. Once that redirect is in place and verified working, we can remove the original first page.