Rename 'console' section of sidebar to 'web console'
The idea here is to be more descriptive and less brand-namey about the console.
Check with marketing to make sure this is OK, but I think it makes sense, esp.
since there are many kinds of consoles out in your infrastructure. I think this
naming makes sense in the sidebar, because it will help people navigate, but is
not necessary elsewhere (i.e. we can keep calling it 'the PE console' in most
places, and should).