Update sample hiera config for Pysch parser
The previous sample hiera config was invalid yaml, which the Syck parser would
coerce into a string. The Pysch parser is less forgiving, and throws an error
like "Failed to start Hiera: Psych::SyntaxError:
(/etc/puppetlabs/puppet/hiera.yaml): found character that cannot start any
token while scanning for the next token at line 6 column 5" when that config is
used. % is a special delimeter in Yaml, so to use it as intended with Hiera, it
must be wrapped in quotes so that Pysch knows it is a string.