Document changes to report format (version 4)
This commit introduces changes for version 4 of the
puppet report format.  This particular commit is related
to the addition of a `containment_path` attribute to the
`resource_status` objects; however, there will probably
be a few additional changes that can be rolled into v4
in subsequent commits prior to the Puppet 3.3.0 release.