Comment out nonfunctional feature -- external facts do not work on windows
Facter does not currently implement external facts support on Windows
as the docs describe. At some point in the future it will probably
actually work as documented so it's not worth removing this and then
having to pull it back out of version control before this commit, but
it shouldn't be advertised as working when it does not.