As the prophets foretold, the user type's feature table changed & broke our references
Remember how I said we'd regret that regex and that providers really needed to
specify their features as data instead of code? Yeah, that.

In Puppet 3.2.0, the user and group types added a `libuser` feature -- not only
did this break our hack to get password management displaying correctly, but it
uses the same bad habit as password management and requires yet another layer of
hackery.

This commit changes our approach for hacking these awful hacks, because I was
unable to comprehend and modify my own godawful regex. The new method ought to
be dumber and cleaner. We also manually fix the three affected reference versions.