(broken) Remove pygments cache plugin
Jekyll version mumble mumble something made syntax highlighting much more
efficient by switching to pygments.rb, which apparently runs a single pygments
process and uses an IO object to shove data through it. The upshot is that we
can get rid of the cache plugin and only suffer about a 10-15 second penalty.
(Remember that this plugin originally took about a minute off our build time.)

Oh, and if we don't delete it, the site refuses to build, lol.