Use ERB to template the documentation for each type
This -- the way to arrange/format the docs for an individual type -- is the
heart of the type reference, and we're doing enough complicated things with the
underlying data that I've concluded we can't make it non-gross. So the right
approach is probably to isolate the grossness as much as possible.

Doing a big chain of string appends mixed with variable assignments is probably
not the way to isolate the grossness! So instead, let's assign a few of the more
expensive variables and then push most of the formatting logic out into a big
nasty ERB template.