Guts: Override in-page TOC for type reference
That TOC was literally about six screens long. On a big monitor. Blow it away
and make a smaller multi-columned one, so you can see the start of the page from
the get-go! This isn't the most sustainable way to do this, but it at least
solves the worst of the fugly.