(#18268) (#21243) (#22368) (#22700) windows installer improvements
Adding minimum version information to the documentation
relating to new and unreleased versions of puppet code.

to the path variable automatically and documents the
first version of the installer that allows that.

account user and how it is specified in the installer.

on 22368 with the puppet agent variable to add the
minimum version.

startup mode for windows and the different items you
can pass to the installer.