Deprecate documentation for PuppetDB 1.3.x and 1.4.x
We have decided to deprecate all older branches now that PE has committed
to 1.5.x and no longer requires 1.3.x to stay open.

This patch pins the older revisions to tags I have created in puppetdb in the
form 1.1.x-docs so that we can close out the older branches for good.

It also puts up a deprecation message on the older 1.3.x and 1.4.x docs and
moves them into the older revisions area of the index.

As a bonus I've added a link to the master version of the documentation from
the index page, as we previously haven't provided that to our users.

Signed-off-by: Ken Barber <ken@bob.sh>