Rename puppet_projects_redmine_workflow
The "Redmine Workflow for Puppet Open-Source Projects" document has been
renamed to reflect the impending migration to JIRA. However, instead of just
swapping "JIRA" in for "Redmine", the title of this document has is now
tracker-neutral: "Workflow for Puppet Open-Source Projects".

This is in case we ever have to do this sort of migration again. All direct
references to this specific document have also been updated.