Console inaccessible on Ubuntu in PE 2.8.3
Added known issue and workaround for the console being inaccessible on
Ubuntu + PE 2.8.3.

This appears to affect upgrades, no only fresh installations

I initially didn't think a known issue was necessary, since most users
are not doing fresh installs of PE 2.8.3, but since it affects upgrades,
this probably needs to be publicly documented for customers.

https://jira.puppetlabs.com/browse/PE-1679