Allow building references with unreleased Puppet versions (VERSION=x.x.x COMMIT=abc0567)
We've run into this annoyance a few times, where we know what's going into the
coming version but we can't generate the references until the release is tagged.
Makes previewing a bit of a pain.

Long story short, we were being authoritarian about validating things without
providing an override. So here's the override: if you provide both VERSION and
COMMIT environment variables, the reference generator will check out the commit
instead of locking you to a tag.