Guts: Add sensible fallback sidebar, remove "legacy" layout
I've meant to do this since forever, but wasn't sure what to put in the fallback
sidebar -- should I put every file into a large meta-document of some kind? How
could I make a default sidebar comprehensive? Then I realized I probably didn't
have to make it comprehensive, and could just link to the eight or so most
popular pages and call it good.

Anyway, this is one of the first commits in a while that *simplifies* our site
structure (by removing a redundant layout), so that's pretty cool.

As a side effect, this moves the sidebar on the front page to the left. Oh well.