PDB-264 Prepare documentation for PuppetDB 1.6.0rc1
This commit prepares the documentation site for the upcoming 1.6.0rc1 release
of PuppetDB.

This slightly differs from our normal approach, in that we normally do not do
an RC release. This time, we are doing all the normal work for a feature
release but modifying the index change so that the pre-release is categorized
accordingly, and the 1.5 documentation stays prominent until we do the final.

When we do the final release, the link for 'latest' and the index update will
be required.

Signed-off-by: Ken Barber <ken@bob.sh>