Parenthesis are mandatory with rvalue-functions
This fixes a problem with the reference documentation where it was not
explained that parenthesis are only optional for statement-level
functions.

Here is an example :

    $a = function "argument" $b = literal

With the "old" parser, this will result in a parse error. With the new
parser, it seems that it is interpreted like this :

    $a = 'function';
    'argument';
    $b = 'literal';

Which might cause puzzling bugs.