Mass update: "any legal log level" was never a legal value for exec's `logoutput`
Making an exception to our usual policy on mass edits to historical references,
because this was actively misleading rather than just vague or non-optimal.

In the update to the embedded doc string, I added more context about the
`loglevel` metaparameter; for the mass update, I did as little work as possible
and just deleted the bad info.