Update for accuracy around split installs.
It's not possible to "install the database support role on the console's
node" except during all-in-one installations.

If it's a split install it's:

  * Master
  * Database Support (Including PuppetDB)
  * Console

The original text leads users to believe that they can install the
database support role (e.g. PostgreSQL + PuppetDB), and then install the
console ON THE SAME NODE.

If you do that, everything asplodes.