Guts: Band-aid the reference generator under Ruby 2.0 and Bundler
'kay, it looks like if you rewrite $RUBYLIB and then execute Ruby in a subshell,
and your initial Ruby process was started under Bundler, it will blow up. That's
a problem because we're relying on that type of shenanigans to sanitize our load
path in the references generator.

So instead of rewriting the whole thing, we'll preserve the first element in it
(which SHOULD be the only one Bundler relies on).

Another fix: It looks like grep doesn't work the same way in Ruby 2, so we'll
just do each_line instead (which is better for getting a list of tags anyway).