Rearrange the 3.5 nav sidebar a bit
I was sleepwalking a bit when arranging these sections. Language should go
before a bunch of that other stuff, references should be higher and not have the
word "generated" in it (because who cares), the type reference may as well have
its own top-level item because it's just that important, and a few sections
should talk about what's included in them so readers know whether they want to
click.