Puppet 3.0-3.4: Change all references links to 3.4.stable
Eric0 noticed that the 3.latest links were pointing to 3.5 now, which is wrong
for this version of the document. This change makes them still a bit fuzzy
(which is part of the reason we switched to .y versioning on these docs in the
first place), but they're pointing to something that's closer to accurate, at
least.