(DOC-717) Fix puppet resource man page
For some reason, having an older version of facter in our Gemfile (and
vendor/bundle/ruby/2.0.0/gems) was blowing up the puppet resource man page (and
ONLY that man page, also FOR SOME REASON). This commit removes facter from the
gemfile (since the team can't remember why it was there in the first place) and
updates the puppet resource man page with appropriate bits.

UNFORTUNATELY, I also noticed that the man page generator gets contamination
from any other Puppets in the load path, so sometime soon I'm going to have to
renovate that to quarantine the process that extracts the text, same as I did
with the type reference.