Break down PE and Communtiy versions of PuppetDB
This changes the index for PuppetDB to differentiate between community and PE
versions for our older documentation. This avoids confusion that I've received
where we are stating for example 1.6.x community is marked as stable, but in
reality this is only true for PE packages.

This also starts us on the road for making PE more of a first class citizen
in the rest of our documentation.

Signed-off-by: Ken Barber <ken@bob.sh>