Fixing the hiera complete example for ntp module changes
At some point the ntp module stopped using booleans in restrict. It now
uses an array. If you want no restrictions you must pass in the empty
string. Leaving behind if thats a good idea or not, the docs are now a
bit better. This was brought to my attention on the mailing list:

https://groups.google.com/forum/?utm_medium=email&utm_source=footer#!msg/puppet-users/KeaFjBNFH78/BXPy77-dtn0J