(DOC-835) "How to change Console Port" lists steps no longer needed
This patch removes two entries from the steps for changing the
Console's port.  These two steps reference items that are no longer
valid for PE 3.2+.

The first item was a reference to `reporturl` in `puppet.conf`. This
is only needed if the `http` report processor is used. The console now
uses a report handler named `console`.

The second item refers to the old ENC script, which no longer exists
and has been replaced with a new node terminus type.